extend sectors tests
