+++ integration_tests/src/tests/mod.rs
@@ -14,7 +14,8 @@ mod datacap_tests;
 pub use datacap_tests::*;
 mod evm_test;
 pub use evm_test::*;
-// mod extend_sectors_test;
+mod extend_sectors_test;
+pub use extend_sectors_test::*;
 // mod init_test;
 // mod market_miner_withdrawal_test;
 // mod multisig_test;
