+++ test_vm/tests/extend_sectors_test.rs
@@ -1,36 +1,24 @@
-use fvm_ipld_bitfield::BitField;
+use fil_actor_miner::{max_prove_commit_duration, PowerPair, Sectors, State as MinerState};
+use fil_actors_integration_tests::tests::{
+    commit_sector_with_max_duration_deal_test, extend,
+    extend_sector_up_to_max_relative_extension_test, extend_updated_sector_with_claims_test,
+};
+use fil_actors_integration_tests::util::{
+    advance_by_deadline_to_epoch, advance_by_deadline_to_epoch_while_proving,
+    advance_by_deadline_to_index, advance_to_proving_deadline, create_accounts, create_miner,
+    cron_tick, expect_invariants, invariant_failure_patterns, market_add_balance,
+    market_publish_deal, miner_precommit_sector, miner_prove_sector, submit_windowed_post,
+    verifreg_add_client, verifreg_add_verifier,
+};
+use fil_actors_runtime::runtime::Policy;
+use fil_actors_runtime::{DealWeight, EPOCHS_IN_DAY};
 use fvm_ipld_blockstore::{Blockstore, MemoryBlockstore};
-use fvm_shared::address::Address;
 use fvm_shared::bigint::Zero;
-use fvm_shared::clock::ChainEpoch;
 use fvm_shared::econ::TokenAmount;
 use fvm_shared::piece::PaddedPieceSize;
 use fvm_shared::sector::{RegisteredSealProof, SectorNumber, StoragePower};
-
-use fil_actor_market::DealMetaArray;
-use fil_actor_market::{SectorDeals, State as MarketState};
-use fil_actor_miner::{
-    max_prove_commit_duration, power_for_sector, ExpirationExtension, ExpirationExtension2,
-    ExtendSectorExpiration2Params, ExtendSectorExpirationParams, Method as MinerMethod, PowerPair,
-    ProveReplicaUpdatesParams2, ReplicaUpdate2, SectorClaim, Sectors, State as MinerState,
-};
-use fil_actor_verifreg::Method as VerifregMethod;
-use fil_actors_runtime::runtime::Policy;
-use fil_actors_runtime::test_utils::{make_piece_cid, make_sealed_cid};
-use fil_actors_runtime::{
-    DealWeight, EPOCHS_IN_DAY, STORAGE_MARKET_ACTOR_ADDR, VERIFIED_REGISTRY_ACTOR_ADDR,
-};
-use test_vm::expects::Expect;
-use test_vm::util::{
-    advance_by_deadline_to_epoch, advance_by_deadline_to_epoch_while_proving,
-    advance_by_deadline_to_index, advance_to_proving_deadline, bf_all, create_accounts,
-    create_miner, cron_tick, expect_invariants, invariant_failure_patterns, market_add_balance,
-    market_publish_deal, miner_precommit_sector, miner_prove_sector, sector_deadline,
-    submit_windowed_post, verifreg_add_client, verifreg_add_verifier,
-};
 use test_vm::TestVM;
-use vm_api::trace::ExpectInvocation;
-use vm_api::util::{apply_ok, get_state, DynBlockstore};
+use vm_api::util::{get_state, DynBlockstore};
 use vm_api::VM;
 
 #[test]
@@ -45,80 +33,6 @@ fn extend2_legacy_sector_with_deals() {
     let store = MemoryBlockstore::new();
     let v = TestVM::<MemoryBlockstore>::new_with_singletons(&store);
     extend_legacy_sector_with_deals_inner(&v, true, &v);
-}
-
-#[allow(clippy::too_many_arguments)]
-fn extend(
-    v: &dyn VM,
-    worker: Address,
-    maddr: Address,
-    deadline_index: u64,
-    partition_index: u64,
-    sector_number: SectorNumber,
-    new_expiration: ChainEpoch,
-    power_delta: PowerPair,
-    v2: bool,
-) {
-    let extension_method = match v2 {
-        false => MinerMethod::ExtendSectorExpiration as u64,
-        true => MinerMethod::ExtendSectorExpiration2 as u64,
-    };
-
-    match v2 {
-        false => {
-            let extension_params = ExtendSectorExpirationParams {
-                extensions: vec![ExpirationExtension {
-                    deadline: deadline_index,
-                    partition: partition_index,
-                    sectors: BitField::try_from_bits([sector_number].iter().copied()).unwrap(),
-                    new_expiration,
-                }],
-            };
-            apply_ok(
-                v,
-                &worker,
-                &maddr,
-                &TokenAmount::zero(),
-                extension_method,
-                Some(extension_params),
-            );
-        }
-        true => {
-            let extension_params = ExtendSectorExpiration2Params {
-                extensions: vec![ExpirationExtension2 {
-                    deadline: deadline_index,
-                    partition: partition_index,
-                    sectors: BitField::try_from_bits([sector_number].iter().copied()).unwrap(),
-                    new_expiration,
-                    sectors_with_claims: vec![],
-                }],
-            };
-            apply_ok(
-                v,
-                &worker,
-                &maddr,
-                &TokenAmount::zero(),
-                extension_method,
-                Some(extension_params),
-            );
-        }
-    };
-
-    let mut expect_invoke =
-        vec![Expect::reward_this_epoch(maddr), Expect::power_current_total(maddr)];
-
-    if !power_delta.is_zero() {
-        expect_invoke.push(Expect::power_update_claim(maddr, power_delta));
-    }
-
-    ExpectInvocation {
-        from: worker,
-        to: maddr,
-        method: extension_method,
-        subinvocs: Some(expect_invoke),
-        ..Default::default()
-    }
-    .matches(v.take_invocations().last().unwrap());
 }
 
 // TODO: remove usage of _v_concrete which is currently required by mutate_state
@@ -345,306 +259,14 @@ fn extend_legacy_sector_with_deals_inner<BS: Blockstore>(
 fn extend_updated_sector_with_claim() {
     let store = MemoryBlockstore::new();
     let v = TestVM::<MemoryBlockstore>::new_with_singletons(&store);
-    let addrs = create_accounts(&v, 3, &TokenAmount::from_whole(10_000));
-    let seal_proof = RegisteredSealProof::StackedDRG32GiBV1P1;
-    let (owner, worker, verifier, verified_client) = (addrs[0], addrs[0], addrs[1], addrs[2]);
-    let sector_number: SectorNumber = 100;
-    let policy = Policy::default();
-
-    // create miner
-    let miner_id = create_miner(
-        &v,
-        &owner,
-        &worker,
-        seal_proof.registered_window_post_proof().unwrap(),
-        &TokenAmount::from_whole(1_000),
-    )
-    .0;
-    let v = v.with_epoch(200);
-
-    //
-    // Precommit, prove and PoSt empty sector (more fully tested in TestCommitPoStFlow)
-    //
-
-    let expiration = v.epoch() + 360 * EPOCHS_IN_DAY;
-
-    miner_precommit_sector(&v, &worker, &miner_id, seal_proof, sector_number, vec![], expiration);
-
-    // advance time by a day and prove the sector
-    let prove_epoch = v.epoch() + EPOCHS_IN_DAY;
-    advance_by_deadline_to_epoch(&v, &miner_id, prove_epoch);
-    miner_prove_sector(&v, &worker, &miner_id, sector_number);
-    // trigger cron to validate the prove commit
-    cron_tick(&v);
-
-    // advance to proving period and submit post
-
-    let (deadline_info, partition_index) =
-        advance_to_proving_deadline(&v, &miner_id, sector_number);
-    let expected_power_delta =
-        PowerPair { raw: StoragePower::from(32u64 << 30), qa: StoragePower::from(32u64 << 30) };
-    submit_windowed_post(
-        &v,
-        &worker,
-        &miner_id,
-        deadline_info,
-        partition_index,
-        Some(expected_power_delta),
-    );
-
-    // move forward one deadline so sector is mutable
-    advance_by_deadline_to_index(
-        &v,
-        &miner_id,
-        deadline_info.index + 1 % policy.wpost_period_deadlines,
-    );
-
-    // Inspect basic sector info
-
-    let miner_state: MinerState = get_state(&v, &miner_id).unwrap();
-    let initial_sector_info = miner_state.get_sector(&store, sector_number).unwrap().unwrap();
-    assert_eq!(expiration, initial_sector_info.expiration);
-    assert!(initial_sector_info.deal_weight.is_zero()); // 0 space time
-    assert!(initial_sector_info.verified_deal_weight.is_zero()); // 0 space time
-                                                                 // publish verified deals
-
-    // register verifier then verified client
-    let datacap = StoragePower::from(32_u128 << 40);
-    verifreg_add_verifier(&v, &verifier, datacap.clone());
-    verifreg_add_client(&v, &verifier, &verified_client, datacap);
-
-    // add market collateral for clients and miner
-    market_add_balance(&v, &verified_client, &verified_client, &TokenAmount::from_whole(3));
-    market_add_balance(&v, &worker, &miner_id, &TokenAmount::from_whole(64));
-
-    // create 1 verified deal for total sector capacity
-    let deal_start = v.epoch() + EPOCHS_IN_DAY;
-    let deal_ids = market_publish_deal(
-        &v,
-        &worker,
-        &verified_client,
-        &miner_id,
-        "deal1".to_string(),
-        PaddedPieceSize(32u64 << 30),
-        true,
-        deal_start,
-        340 * EPOCHS_IN_DAY,
-    )
-    .ids;
-
-    // replica update
-    let new_cid = make_sealed_cid(b"replica1");
-    let (d_idx, p_idx) = sector_deadline(&v, &miner_id, sector_number);
-    let replica_update = ReplicaUpdate2 {
-        sector_number,
-        deadline: d_idx,
-        partition: p_idx,
-        new_sealed_cid: new_cid,
-        deals: deal_ids.clone(),
-        update_proof_type: fvm_shared::sector::RegisteredUpdateProof::StackedDRG32GiBV1,
-        replica_proof: vec![],
-        new_unsealed_cid: make_piece_cid(b"unsealed from itest vm"),
-    };
-    let updated_sectors: BitField = apply_ok(
-        &v,
-        &worker,
-        &miner_id,
-        &TokenAmount::zero(),
-        MinerMethod::ProveReplicaUpdates2 as u64,
-        Some(ProveReplicaUpdatesParams2 { updates: vec![replica_update] }),
-    )
-    .deserialize()
-    .unwrap();
-    assert_eq!(vec![sector_number], bf_all(updated_sectors));
-
-    let old_power = power_for_sector(seal_proof.sector_size().unwrap(), &initial_sector_info);
-    // check for the expected subcalls
-    ExpectInvocation {
-        from: worker,
-        to: miner_id,
-        method: MinerMethod::ProveReplicaUpdates2 as u64,
-        subinvocs: Some(vec![
-            Expect::market_activate_deals(
-                miner_id,
-                deal_ids.clone(),
-                initial_sector_info.expiration,
-                initial_sector_info.seal_proof,
-            ),
-            ExpectInvocation {
-                from: miner_id,
-                to: VERIFIED_REGISTRY_ACTOR_ADDR,
-                method: VerifregMethod::ClaimAllocations as u64,
-                ..Default::default()
-            },
-            Expect::market_verify_deals(
-                miner_id,
-                vec![SectorDeals {
-                    sector_type: seal_proof,
-                    sector_expiry: initial_sector_info.expiration,
-                    deal_ids: deal_ids.clone(),
-                }],
-            ),
-            Expect::reward_this_epoch(miner_id),
-            Expect::power_current_total(miner_id),
-            Expect::power_update_pledge(miner_id, None),
-            Expect::power_update_claim(
-                miner_id,
-                PowerPair { raw: StoragePower::zero(), qa: 9 * old_power.qa },
-            ),
-        ]),
-        ..Default::default()
-    }
-    .matches(v.take_invocations().last().unwrap());
-
-    // inspect sector info
-
-    let miner_state: MinerState = get_state(&v, &miner_id).unwrap();
-    let sector_info_after_update = miner_state.get_sector(&store, sector_number).unwrap().unwrap();
-    assert_eq!(StoragePower::zero(), sector_info_after_update.deal_weight); // 0 space time
-
-    assert_eq!(
-        DealWeight::from((sector_info_after_update.expiration - v.epoch()) * (32i64 << 30)),
-        sector_info_after_update.verified_deal_weight
-    ); // 32 GiB * the remaining life of the sector
-
-    // extend the updated sector
-
-    // For clarity in checking power_base_epoch, we increment epoch by 1
-    let curr_epoch = v.epoch();
-    let v = v.with_epoch(curr_epoch + 1);
-
-    let market_state: MarketState = get_state(&v, &STORAGE_MARKET_ACTOR_ADDR).unwrap();
-    let deal_states = DealMetaArray::load(&market_state.states, v.store).unwrap();
-    let deal_state = deal_states.get(deal_ids[0]).unwrap().unwrap();
-    let claim_id = deal_state.verified_claim;
-
-    let extension_params = ExtendSectorExpiration2Params {
-        extensions: vec![ExpirationExtension2 {
-            deadline: d_idx,
-            partition: partition_index,
-            sectors: BitField::new(),
-            new_expiration: sector_info_after_update.expiration + 60 * EPOCHS_IN_DAY,
-            sectors_with_claims: vec![SectorClaim {
-                sector_number,
-                maintain_claims: vec![claim_id],
-                drop_claims: vec![],
-            }],
-        }],
-    };
-    apply_ok(
-        &v,
-        &worker,
-        &miner_id,
-        &TokenAmount::zero(),
-        MinerMethod::ExtendSectorExpiration2 as u64,
-        Some(extension_params),
-    );
-
-    let miner_state: MinerState = get_state(&v, &miner_id).unwrap();
-    let sector_info_after_extension =
-        miner_state.get_sector(&store, sector_number).unwrap().unwrap();
-    assert_eq!(StoragePower::zero(), sector_info_after_extension.deal_weight); // 0 space time
-
-    assert_eq!(
-        DealWeight::from((sector_info_after_extension.expiration - v.epoch()) * (32i64 << 30)),
-        sector_info_after_extension.verified_deal_weight
-    ); // 32 GiB * the remaining life of the sector
-
-    assert_eq!(sector_info_after_extension.power_base_epoch, v.epoch());
-    assert_eq!(sector_info_after_update.activation, sector_info_after_extension.activation);
-    assert_eq!(
-        sector_info_after_extension.replaced_day_reward,
-        sector_info_after_update.expected_day_reward
-    );
+    extend_updated_sector_with_claims_test(&v);
 }
 
 #[test]
 fn extend_sector_up_to_max_relative_extension() {
     let store = MemoryBlockstore::new();
     let v = TestVM::<MemoryBlockstore>::new_with_singletons(&store);
-    let addrs = create_accounts(&v, 3, &TokenAmount::from_whole(10_000));
-    let seal_proof = RegisteredSealProof::StackedDRG32GiBV1P1;
-    let (owner, worker) = (addrs[0], addrs[0]);
-    let sector_number: SectorNumber = 100;
-    let policy = Policy::default();
-
-    // create miner
-    let miner_id = create_miner(
-        &v,
-        &owner,
-        &worker,
-        seal_proof.registered_window_post_proof().unwrap(),
-        &TokenAmount::from_whole(1_000),
-    )
-    .0;
-    let v = v.with_epoch(200);
-
-    //
-    // Precommit, prove and PoSt empty sector (more fully tested in TestCommitPoStFlow)
-    //
-    let sector_start =
-        v.epoch() + max_prove_commit_duration(&Policy::default(), seal_proof).unwrap();
-
-    miner_precommit_sector(
-        &v,
-        &worker,
-        &miner_id,
-        seal_proof,
-        sector_number,
-        vec![],
-        sector_start + 180 * EPOCHS_IN_DAY,
-    );
-
-    // advance time to max seal duration and prove the sector
-    advance_by_deadline_to_epoch(&v, &miner_id, sector_start);
-    miner_prove_sector(&v, &worker, &miner_id, sector_number);
-    // trigger cron to validate the prove commit
-    cron_tick(&v);
-
-    // inspect sector info
-    let mut miner_state: MinerState = get_state(&v, &miner_id).unwrap();
-    let mut sector_info = miner_state.get_sector(&store, sector_number).unwrap().unwrap();
-    assert_eq!(180 * EPOCHS_IN_DAY, sector_info.expiration - sector_info.activation);
-
-    // advance to proving period and submit post
-    let (deadline_info, partition_index) =
-        advance_to_proving_deadline(&v, &miner_id, sector_number);
-    let expected_power_delta =
-        PowerPair { raw: StoragePower::from(32u64 << 30), qa: StoragePower::from(32u64 << 30) };
-    submit_windowed_post(
-        &v,
-        &worker,
-        &miner_id,
-        deadline_info,
-        partition_index,
-        Some(expected_power_delta),
-    );
-
-    // move forward one deadline so advanceWhileProving doesn't fail double submitting posts
-    advance_by_deadline_to_index(
-        &v,
-        &miner_id,
-        deadline_info.index + 1 % policy.wpost_period_deadlines,
-    );
-
-    // Extend the sector by the max relative extension.
-    let new_expiration = v.epoch() + policy.max_sector_expiration_extension;
-    let expected_power_delta = PowerPair::zero();
-    extend(
-        &v,
-        worker,
-        miner_id,
-        deadline_info.index,
-        partition_index,
-        sector_number,
-        new_expiration,
-        expected_power_delta,
-        false,
-    );
-
-    miner_state = get_state(&v, &miner_id).unwrap();
-    sector_info = miner_state.get_sector(&store, sector_number).unwrap().unwrap();
-    assert_eq!(policy.max_sector_expiration_extension, sector_info.expiration - v.epoch());
+    extend_sector_up_to_max_relative_extension_test(&v);
 }
 
 #[test]
@@ -652,94 +274,4 @@ fn commit_sector_with_max_duration_deal() {
     let store = MemoryBlockstore::new();
     let v = TestVM::<MemoryBlockstore>::new_with_singletons(&store);
     commit_sector_with_max_duration_deal_test(&v);
-}
-
-fn commit_sector_with_max_duration_deal_test(v: &dyn VM) {
-    let addrs = create_accounts(v, 3, &TokenAmount::from_whole(10_000));
-    let seal_proof = RegisteredSealProof::StackedDRG32GiBV1P1;
-    let (owner, worker, verifier, verified_client) = (addrs[0], addrs[0], addrs[1], addrs[2]);
-    let sector_number: SectorNumber = 100;
-    let policy = Policy::default();
-
-    // create miner
-    let miner_id = create_miner(
-        v,
-        &owner,
-        &worker,
-        seal_proof.registered_window_post_proof().unwrap(),
-        &TokenAmount::from_whole(1_000),
-    )
-    .0;
-    v.set_epoch(200);
-
-    //
-    // publish verified deals
-    //
-
-    // register verifier then verified client
-    let datacap = StoragePower::from(32_u128 << 40);
-    verifreg_add_verifier(v, &verifier, datacap.clone());
-    verifreg_add_client(v, &verifier, &verified_client, datacap);
-
-    // add market collateral for clients and miner
-    market_add_balance(v, &verified_client, &verified_client, &TokenAmount::from_whole(3));
-    market_add_balance(v, &worker, &miner_id, &TokenAmount::from_whole(64));
-
-    let deal_lifetime = policy.max_sector_expiration_extension
-        - max_prove_commit_duration(&policy, seal_proof).unwrap_or_default();
-    // create 1 verified deal for total sector capacity for 6 months
-    let deal_start = v.epoch() + max_prove_commit_duration(&Policy::default(), seal_proof).unwrap();
-    let deals = market_publish_deal(
-        v,
-        &worker,
-        &verified_client,
-        &miner_id,
-        "deal1".to_string(),
-        PaddedPieceSize(32u64 << 30),
-        true,
-        deal_start,
-        deal_lifetime,
-    )
-    .ids;
-
-    //
-    // Precommit, prove and PoSt empty sector (more fully tested in TestCommitPoStFlow)
-    //
-    miner_precommit_sector(
-        v,
-        &worker,
-        &miner_id,
-        seal_proof,
-        sector_number,
-        deals,
-        deal_start + deal_lifetime,
-    );
-
-    // advance time to max seal duration and prove the sector
-    advance_by_deadline_to_epoch(v, &miner_id, deal_start);
-    miner_prove_sector(v, &worker, &miner_id, sector_number);
-    // trigger cron to validate the prove commit
-    cron_tick(v);
-
-    // advance to proving period and submit post
-    let (deadline_info, partition_index) = advance_to_proving_deadline(v, &miner_id, sector_number);
-    let expected_power_delta = PowerPair {
-        raw: StoragePower::from(32u64 << 30),
-        qa: 10 * StoragePower::from(32u64 << 30),
-    };
-    submit_windowed_post(
-        v,
-        &worker,
-        &miner_id,
-        deadline_info,
-        partition_index,
-        Some(expected_power_delta),
-    );
-    // inspect sector info
-    let miner_state: MinerState = get_state(v, &miner_id).unwrap();
-    let sector_info = miner_state
-        .get_sector(&DynBlockstore::wrap(v.blockstore()), sector_number)
-        .unwrap()
-        .unwrap();
-    assert_eq!(deal_lifetime, sector_info.expiration - sector_info.activation);
 }
